Collision work
- added collision rects to delivery robot and three shot flyer
- changed the collision rect to use s16
- disabled particle effects and item drop because they're current buggy

+++ project/source/engine/gameobject_template_types.h
@@ -15,10 +15,10 @@ typedef struct gameobject_template
 	s8 startHealth;
 	s8 damage;
 
-	s8 rectLeft;
-	s8 rectTop;
-	s8 rectRight;
-	s8 rectBottom;
+	s16 rectLeft;
+	s16 rectTop;
+	s16 rectRight;
+	s16 rectBottom;
 
 	u8 objectType;
 	const ResourceInfo* resourceInfo;
