+++ project/source/engine/object_types.h
@@ -54,10 +54,10 @@
 	\
 	s8 health;\
 	s8 damage;\
-	s8 rectLeft;\
-	s8 rectTop;\
-	s8 rectRight;\
-	s8 rectBottom;\
+	s16 rectLeft;\
+	s16 rectTop;\
+	s16 rectRight;\
+	s16 rectBottom;\
 	u8 objectType; \
 	const ResourceInfo* resourceInfo;\
 	const ResourceInfo** extraResources;\
